refactor(api): extract service layer from sessions.py (1399 to 769 lines)
Phase 3: turn the fat sessions router into a thin adapter over a new
api/services/ layer. Pure code movement — no endpoint path, status,
response model, or SQL change.

New api/services/ modules:
- model_loading.py: _load_obml, _model_load_fields
- session_lifecycle.py: _get_store, _session_response
- query_compilation.py: _resolve_dialect, _obsql_translation_errors,
  _join_path_steps + compile_query_or_raise / build_compile_response /
  compile_query_for_plan / build_semantic_ql_compile_response
- query_execution.py: _build_execute_response, _run_with_cache, TTL/cache
  metadata helpers, _build_cached_response, _physical_tables_for

The big compile/execute/plan/semantic-ql handlers are now thin: resolve
deps, call the service, translate domain errors to HTTPException. Moved
names are re-imported into the sessions namespace (with __all__) so the
shortcuts / oneshot / model_api importers resolve unchanged.

Add direct service unit tests (no ASGI). Full suite green (2479 passed,
167 skipped); ruff/format/mypy clean; 0 import cycles.
